\name{classify_model}
\alias{classify_model}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
makes predictions from a train_model() object.
}
\description{
Uses a trained model from the train_model() function to classify new data.
}
\usage{
classify_model(corpus, model, s = 0.01, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{corpus}{
Class of type matrix_container generated by the create_corpus() function.
}
  \item{model}{
Slot for trained SVM, Naive Bayes, boosting, bagging, RandomForests, glmnet, decision tree, neural network, or maximum entropy model generated by train_model().
}
  \item{s}{
Penalty parameter lambda for glmnet classification.
}
  \item{dots}{
Additional parameters to be passed into the predict() function of any algorithm.
}
}
\details{
Only one model may be passed in at a time for classification. See train_models() and classify_models() to train and classify using multiple algorithms.
}
\value{
Returns a data frame of predicted codes and probabilities for the specified algorithm.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Timothy P. Jurka <tpjurka@ucdavis.edu>, Loren Collingwood <lorenc2@uw.edu>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
library(RTextTools)
data <- read_data(system.file("data/NYTimes.csv.gz",package="RTextTools"),type="csv")
data <- data[sample(1:3100,size=1000,replace=FALSE),]
matrix <- create_matrix(cbind(data$Title,data$Subject), language="english", 
removeNumbers=TRUE, stemWords=FALSE, weighting=weightTfIdf)
corpus <- create_corpus(matrix,data$Topic.Code,trainSize=1:750, testSize=751:1000, 
virgin=FALSE)
glmnet_model <- train_model(corpus,"GLMNET")
maxent_model <- train_model(corpus,"MAXENT")
svm_model <- train_model(corpus,"SVM")
glmnet_results <- classify_model(corpus,glmnet_model)
maxent_results <- classify_model(corpus,maxent_model)
svm_results <- classify_model(corpus,svm_model)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}