\name{create_analytics}
\alias{create_analytics}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
creates an object of class analytics given classification results.
}
\description{
Takes the results from functions classify_model() or classify_models() and computes various statistics to help interpret the data.
}
\usage{
create_analytics(corpus, classification_results, b=1, threshold=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{corpus}{
Class of type matrix_container generated by the create_corpus() function.
}
  \item{classification_results}{
A cbind() of result objects returned by classify_model(), or the object returned by classify_models().
}
  \item{b}{
b-value for generating precision, recall, and F-scores statistics.
}
  \item{threshold}{
The number of algorithms greater than or equal to this threshold that agree on the same topic. (e.g. a threshold value of 3 will search for those documents where 3 or more algorithms agreed)
}
}
\details{
Object of class analytics has four slots: algorithm_summary, ensemble_summary, document_summary, and label_summary. They can be accessed using the @ operator (e.g. analytics@algorithm_summary).
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Timothy P. Jurka <tpjurka@ucdavis.edu>, Loren Collingwood <lorenc2@uw.edu>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
library(RTextTools)
data <- read_data(system.file("data/NYTimes.csv.gz",package="RTextTools"),type="csv")
data <- data[sample(1:3100,size=1000,replace=FALSE),]
matrix <- create_matrix(cbind(data$Title,data$Subject), language="english", 
removeNumbers=TRUE, stemWords=FALSE, weighting=weightTfIdf)
corpus <- create_corpus(matrix,data$Topic.Code,trainSize=1:750, testSize=751:1000, 
virgin=FALSE)
models <- train_models(corpus, algorithms=c("GLMNET","MAXENT","SVM"))
results <- classify_models(corpus, models)
analytics <- create_analytics(corpus, results)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}