\name{create_ensembleSummary}
\alias{create_ensembleSummary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
creates a summary with ensemble coverage and precision.
}
\description{
Creates a summary with ensemble coverage and precision values for an ensemble greater than the threshold specified.
}
\usage{
create_ensembleSummary(score_summary, threshold)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{score_summary}{
The score_summary slot from the analytics generated by create_analytics().
}
  \item{threshold}{
The number of algorithms greater than or equal to this threshold that agree on the same topic. (e.g. a threshold value of 3 will search for those documents where 3 or more algorithms agreed)
}
}
\details{
This summary is created in the create_analytics function.  Note that a threshold value of 3 will return ensemble coverage and precision statistics for topic codes that had 3 or more (i.e. >=3) algorithms agree on the same topic code.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Loren Collingwood <lorenc2@uw.edu>, Timothy P. Jurka <tpjurka@ucdavis.edu>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
library(RTextTools)
data <- read_data(system.file("data/NYTimes.csv.gz",package="RTextTools"),type="csv")
data <- data[sample(1:3100,size=1000,replace=FALSE),]
matrix <- create_matrix(cbind(data$Title,data$Subject), language="english", 
removeNumbers=TRUE, stemWords=FALSE, weighting=weightTfIdf)
corpus <- create_corpus(matrix,data$Topic.Code,trainSize=1:750, testSize=751:1000, 
virgin=FALSE)
models <- train_models(corpus, algorithms=c("GLMNET","MAXENT","SVM"))
results <- classify_models(corpus, models)
analytics <- create_analytics(corpus, results)
ensemble <- create_ensembleSummary(analytics@document_summary,3)
ensemble
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}