\name{print_algorithms}
\alias{print_algorithms}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
prints available algorithms for train_model() and train_models().
}
\description{
An informative function that displays options for the \code{algorithms} parameter in \code{\link{train_model}} and \code{\link{train_models}}.
}
\usage{
print_algorithms()
}
%- maybe also 'usage' for other objects documented here.

\value{
Prints a list of available algorithms.
}
\author{
Timothy P. Jurka <tpjurka@ucdavis.edu>
}
\examples{
library(RTextTools)
print_algorithms()
}
\keyword{method}
