\name{summary.analytics}
\alias{summary.analytics}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
summarizes the \code{\link{analytics-class}} class
}
\description{
Returns a summary of the contents within an object of class \code{\link{analytics-class}}.
}
\usage{
\method{summary}{analytics}(object, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class \code{\link{analytics-class}} containing the output of the \code{\link{create_analytics}} function.
}
  \item{\dots}{
Additional parameters to be passed onto the summary function.
}
}

\author{
Timothy P. Jurka
}
\keyword{method}
